+++ packages/lumx-react/src/components/list/List.tsx
@@ -7,7 +7,7 @@ import { GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/
 import { mergeRefs } from '@lumx/react/utils/mergeRefs';
 
 import classNames from 'classnames';
-import React, { Key, ReactNode, Ref, useRef } from 'react';
+import React, { Key, ReactNode, Ref, SyntheticEvent, useRef } from 'react';
 import { useInteractiveList } from './useInteractiveList';
 
 /**
@@ -27,8 +27,14 @@ interface ListProps extends GenericProps {
     listElementRef?: Ref<HTMLUListElement>;
     /** Whether custom colors are applied to this component or not. */
     useCustomColors?: boolean;
-    /** The function called when a list item is selected. */
-    onListItemSelected?(key: Key, index: number): void;
+    /**
+     * Callback used to retrieved the select entry.
+     *
+     * @param key   React key of the selected item.
+     * @param index Index of the selected item among the sibling items.
+     * @param evt   Source event (either mouse or keyboard event).
+     */
+    onListItemSelected?(key: Key, index: number, evt: SyntheticEvent): void;
 }
 
 /**